requirements.txt
altair                    5.5.0
attrs                     25.3.0
beautifulsoup4            4.13.4
blinker                   1.9.0
cachetools                6.1.0
certifi                   2025.7.14
charset-normalizer        3.4.2
click                     8.2.1
clip                      1.0
colorama                  0.4.6
ddgs                      9.5.1
decorator                 5.2.1
filelock                  3.18.0
fsspec                    2025.7.0
ftfy                      6.3.1
gitdb                     4.0.12
GitPython                 3.1.45
idna                      3.10
imageio                   2.37.0
imageio-ffmpeg            0.6.0
Jinja2                    3.1.6
joblib                    1.5.1
jsonschema                4.25.0
jsonschema-specifications 2025.4.1
kagglehub                 0.3.12
llvmlite                  0.44.0
lxml                      6.0.0
MarkupSafe                3.0.2
more-itertools            10.7.0
moviepy                   2.2.1
mpmath                    1.3.0
narwhals                  2.0.0
networkx                  3.5
numba                     0.61.2
numpy                     2.2.0
openai-whisper            20250625
opencv-python             4.12.0.88
packaging                 25.0
pandas                    2.3.1
pillow                    11.3.0
pip                       23.1.2
primp                     0.15.0
proglog                   0.1.12
protobuf                  6.31.1
pyarrow                   21.0.0
pydeck                    0.9.1
python-dateutil           2.9.0.post0
python-dotenv             1.1.1
pytz                      2025.2
PyYAML                    6.0.2
referencing               0.36.2
regex                     2025.7.34
requests                  2.32.4
rpds-py                   0.26.0
scikit-learn              1.7.1
scipy                     1.16.1
setuptools                65.5.0
six                       1.17.0
smmap                     5.0.2
soupsieve                 2.7
streamlit                 1.47.1
sympy                     1.14.0
tenacity                  9.1.2
threadpoolctl             3.6.0
tiktoken                  0.9.0
toml                      0.10.2
torch                     2.7.1
torchvision               0.22.1
tornado                   6.5.1
tqdm                      4.67.1
typing_extensions         4.14.1
tzdata                    2025.2
urllib3                   2.5.0
watchdog                  6.0.0
wcwidth                   0.2.13